Refactor authentication related code into separate component
